+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Scripts.hs
@@ -130,13 +130,13 @@ invalidBeforeL = lens invalidBefore (\vi before -> vi {invalidBefore = before})
 invalidHereAfterL :: Lens' ValidityInterval (StrictMaybe SlotNo)
 invalidHereAfterL = lens invalidHereafter (\vi hereAfter -> vi {invalidHereafter = hereAfter})
 
-encodeVI :: ValidityInterval -> Encode ('Closed 'Dense) ValidityInterval
+encodeVI :: ValidityInterval -> Encode (Closed Dense) ValidityInterval
 encodeVI (ValidityInterval f t) = Rec ValidityInterval !> To f !> To t
 
 instance EncCBOR ValidityInterval where
   encCBOR vi = encode (encodeVI vi)
 
-decodeVI :: Decode ('Closed 'Dense) ValidityInterval
+decodeVI :: Decode (Closed Dense) ValidityInterval
 decodeVI = RecD ValidityInterval <! From <! From
 
 instance DecCBOR ValidityInterval where
@@ -213,7 +213,7 @@ instance Era era => EncCBOR (TimelockRaw era) where
 instance Era era => DecCBOR (Annotator (TimelockRaw era)) where
   decCBOR = decode (Summands "TimelockRaw" decRaw)
     where
-      decRaw :: Word -> Decode 'Open (Annotator (TimelockRaw era))
+      decRaw :: Word -> Decode Open (Annotator (TimelockRaw era))
       decRaw 0 = Ann (SumD TimelockSignature <! From)
       decRaw 1 = Ann (SumD TimelockAllOf) <*! D (sequence <$> decCBOR)
       decRaw 2 = Ann (SumD TimelockAnyOf) <*! D (sequence <$> decCBOR)