+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxBody.hs
@@ -167,7 +167,7 @@ boxBody n = invalidField n
 --   The key order looks strange but was choosen for backward compatibility.
 txSparse ::
   EraTxBody era =>
-  ShelleyTxBodyRaw t era -> Encode ('Closed 'Sparse) (ShelleyTxBodyRaw t era)
+  ShelleyTxBodyRaw t era -> Encode (Closed Sparse) (ShelleyTxBodyRaw t era)
 txSparse (ShelleyTxBodyRaw input output cert wdrl fee ttl update hash) =
   Keyed (\i o f t c w u h -> ShelleyTxBodyRaw i o c w f t u h)
     !> Key 0 (To input) -- We don't have to send these in ShelleyTxBodyRaw order