+++ libs/cardano-ledger-core/src/Cardano/Ledger/BaseTypes.hs
@@ -15,10 +15,10 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeData #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE ViewPatterns #-}
-{-# LANGUAGE TypeData #-}
 
 module Cardano.Ledger.BaseTypes (
   module Slotting,